scripts/requirements_validation.txt
tableauserverclient>=0.25
requests>=2.31
python-dotenv>=1.0
